backend/requirements.txt
 
fastapi==0.111.0
uvicorn==0.30.1
torch==2.3.1
transformers==4.42.4
sentence-transformers==3.0.1
scikit-learn==1.5.1
pytesseract==0.3.10  # OCR için, Tesseract'i ayrı kurman gerekecek
pillow==10.4.0  # Resim işleme için
python-multipart==0.0.9  # Dosya upload için